+++ test/src/test_view.cpp
@@ -52,15 +52,14 @@ TEST(test_view, api)
 
     auto metric3 = metrics.initialize_optional<abacus::enum8>(name3);
 
-    std::vector<uint8_t> meta_data(metrics.metadata_bytes());
     std::vector<uint8_t> value_data(metrics.value_bytes());
-    std::memcpy(meta_data.data(), metrics.metadata_data(),
-                metrics.metadata_bytes());
-    std::memcpy(value_data.data(), metrics.value_data(), metrics.value_bytes());
+    std::memcpy(value_data.data(),
+                ((const abacus::metrics&)metrics).value_data(),
+                metrics.value_bytes());
 
     abacus::view view;
 
-    bool success = view.set_meta_data(meta_data.data(), meta_data.size());
+    bool success = view.set_meta_data(metrics.metadata());
     ASSERT_TRUE(success);
     EXPECT_EQ(view.metadata().protocol_version(), abacus::protocol_version());
 
@@ -90,7 +89,8 @@ TEST(test_view, api)
     metric3 = 2;
 
     // and provide new value data to the view
-    success = view.set_value_data(metrics.value_data(), metrics.value_bytes());
+    success = view.set_value_data(
+        ((const abacus::metrics&)metrics).value_data(), metrics.value_bytes());
     ASSERT_TRUE(success);
     view_value0 = view.value<abacus::uint64>(name0);